+++ siepic_forge/__init__.py
@@ -1,4 +1,4 @@
 from .technology import ebeam
 from .component import component, component_names
 
-__version__ = "1.0.0"
+__version__ = "1.1.0dev"