+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.1.0 - Unreleased
+
+### Added
+- Electrical interfaces.
+- TM port specification for 1310 nm with 350 nm width.
+
+
 ## v1.0.0 - 2024-02-13
 
 ### Added